Seperate type definition of request handler from Middleware.
