+++ src/app/http/RequestHandler.ts
@@ -1,0 +1,15 @@
+import { Request, Response, NextFunction } from 'express'
+
+export type SimpleHandler = (req: Request, res: Response) => void
+export type NextHandler = (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => void
+export type ErrorHandler = (
+  err: any,
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => void
+export type RequestHandler = SimpleHandler | NextHandler | ErrorHandler
