+++ src/app/http/middleware/Middleware.ts
@@ -1,19 +1,5 @@
-import { Request, Response, NextFunction } from 'express'
-
-export type SimpleHandler = (req: Request, res: Response) => void
-export type NextHandler = (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => void
-export type ErrorHandler = (
-  err: any,
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => void
-export type Handler = SimpleHandler | NextHandler | ErrorHandler
+import { RequestHandler } from 'Http/RequestHandler'
 
 export default interface Middleware {
-  handler(): Handler
+  handler(): RequestHandler
 }
